#155 adjusts layout
